creating user.php file
this file insert the data into DB.
get name for matches.
#########
__DIR__ (magical constant). The directory of the file. If used inside an include, the directory of the included file
is returned. This is equivalent to dirname(__FILE__). This directory name does not have a trailing slash unless it is the root directory.
##########
a class called Database was create in the connection.php file. Now we use a construct to pass parameters.